+++ lib/fluent_result.dart
@@ -1,7 +1,6 @@
+export 'src/combinators.dart';
 export 'src/errors/errors.dart';
-export 'src/extensions.dart';
 export 'src/methods.dart';
 export 'src/result.dart';
 export 'src/result_config.dart';
 export 'src/result_matcher.dart';
-export 'src/result_of.dart';