+++ lib/src/errors/errors.dart
@@ -1,2 +1,1 @@
-export 'exception_error.dart';
 export 'result_error.dart';